AutoPT/requirements.txt
beautifulsoup4==4.12.3
ipython==8.12.3
jsonlines==4.0.0
langchain==0.2.15
langchain_community==0.2.15
langchain_core==0.2.37
langchain_nvidia_ai_endpoints==0.2.2
langchain_openai==0.1.23
langgraph==0.2.16
nest_asyncio==1.6.0
paramiko==3.4.0
PyYAML==6.0.1
Requests==2.32.3
termcolor==2.4.0
